Simplify Balancer 1st deployment : make it part of FIP-33
Also, only move the 200,000 BAL, and don't create new pools etc
