+++ proposals/description/fip_33b.ts
@@ -1,14 +1,14 @@
 import { ProposalDescription } from '@custom-types/types';
 
-const fip_999: ProposalDescription = {
-  title: 'FIP-XYZ: ------------------------',
+const fip_33b: ProposalDescription = {
+  title: 'FIP-33b: Deposit 200,000 BAL in the BAL/WETH pool',
   commands: [
     {
       target: 'aaveEthPCVDeposit',
       values: '0',
       method: 'withdraw(address,uint256)',
-      arguments: ['{balancerDepositBalWeth}', '200000000000000000000'],
-      description: 'Withdraw 200 WETH from Aave to the BAL/WETH PCVDeposit'
+      arguments: ['{balancerDepositBalWeth}', '250000000000000000000'],
+      description: 'Withdraw 250 WETH from Aave to the BAL/WETH PCVDeposit'
     },
     {
       target: 'bal',
@@ -35,14 +35,16 @@ const fip_999: ProposalDescription = {
   description: `
 
 Summary:
-------------------------
+This proposal is a follow-up to FIP-33: Swap between Balancer DAO and Fei DAO
 
 Specification:
-------------------------
+From the original FIP-33 snapshot: "Fei DAO will deposit BAL into BAL/WETH 80/20 paired with ETH from the PCV and commits to holding for the long run".
 
-Forum discussion: ------------------------
-Snapshot: ------------------------
+This proposal moves the 200,000 BAL received by OTC to the BAL/WETH 80/20 pool, along with 250 WETH from Aave to reduce slippage.
+
+Forum discussion: https://tribe.fei.money/t/fip-33-swap-between-balancer-dao-and-fei-dao/3555
+Snapshot: https://snapshot.org/#/fei.eth/proposal/QmZRQGzAxYFQDTPoXKwsjySMTV3iyT9S8V8zLyscWJGeUq
 `
 };
 
-export default fip_999;
+export default fip_33b;
